Fix loading on some apps
For some reason, our 10 year old app needed the autoload_once_paths
specified in order to autoload constants from those paths. New Rails
applications worked without this. Both are running Zeitwerk on Rails
7.0.4.2, so not sure what the discrepancy is. This was inspired/solved
by looking through ActionText, which does the same thing.
